% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_legacy_api.R
\name{clinical_trial_downloader_two_terms}
\alias{clinical_trial_downloader_two_terms}
\title{A function to download trials for a set of diseases and drugs}
\usage{
clinical_trial_downloader_two_terms(
  condition_terms = NULL,
  treatment_terms = NULL,
  folder,
  url_suffix = NULL
)
}
\arguments{
\item{condition_terms}{Optional. A character vector of conditions (diseases).}

\item{treatment_terms}{Optional. A character vector of interventions (treatments).}

\item{folder}{A full path to store files.}

\item{url_suffix}{Optional. Additional parameters from advanced search starting with "&" (such as "&city=Stockholm").}
}
\value{
Data.frame of clinical trials or a message string.
}
\description{
`clinical_trial_downloader_two_terms()` This function is used to download trials for a set of diseases and drugs
just as they would have been seen during manual search on the website. The search is performed for all combinations of treatments
and diseases.
}
\details{
All terms must be URL compatible. "url_suffix" parameters could be found in the advanced search for
ClinicalTrials.gov (https://clinicaltrials.gov/ct2/search/advanced?cond=&term=&cntry=&state=&city=&dist=). This
function depends on the legacy ClinicalTrials.gov API that may get defunct at any time. The search automatically
applies synonyms for drugs and conditions. In the download folder, one could find a reporting file reporting_file.csv
for the information about the downloaded trials and searches.
}
\examples{

clinical_trial_downloader_two_terms(
condition_terms = c("brain cancer", "glioma"),
treatment_terms = c("cisplatin", "temozolomide", "olaparib"),
folder = "test_downloads",
url_suffix = "&cntry=GB&state=&city=London")
}
