% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_v2_api.R
\name{clinical_trial_downloaderV2_API}
\alias{clinical_trial_downloaderV2_API}
\title{Download all clinical trials for specified conditions, treatment, and parameters via APIv2 ULR}
\usage{
clinical_trial_downloaderV2_API(
  condition_terms = NULL,
  treatment_terms = NULL,
  folder,
  format_suffix = "format=csv",
  url_suffix = NULL
)
}
\arguments{
\item{condition_terms}{A vector of conditions to look for}

\item{treatment_terms}{A vector of therapies to look for}

\item{folder}{A folder to store download files (used only once)}

\item{format_suffix}{Currently supports only CSV. Do not change!}

\item{url_suffix}{Additional API provided provided as URL fragment as in https://clinicaltrials.gov/data-api/api}
}
\value{
Nothing. The function downloads data to a specified folder.
Properties:
Specified conditions, treatments, and suffix could be "".
If the number of trials was 0 for some URLs, saves a file ZERO_URLS.txt with
these URLs.
If some URLs resulted in an error, saves a file FAILED_URLS.txt with
these URLs.
}
\description{
`clinical_trial_downloaderV2_API()` is used to download trials for specified conditions, treatment,
and parameters via APIv2 ULR(see more: https://clinicaltrials.gov/data-api/api). Currently
it can download 1000 trials per request. If resulting number of trials > 1000,
more than one request URL will be generated internally
}
\details{
Currently supports only .csv format of downloads.
Additional download parameters could be provided via url_suffix
The function looks for all combinations of specified treatment and
condition terms
}
\examples{

clinical_trial_downloaderV2_API(condition_terms = c("brain cancer", "glioma", "lung cancer"),
                               treatment_terms = c("cisplatin", "temozolomide", "olaparib"),
                               folder = "test_downloads3",
                               url_suffix = "&query.patient=AREA[LocationCity] London")
results = import_ct_results_v2_csv("test_downloads3")
}
