% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_v2_api.R
\name{download_v2_url}
\alias{download_v2_url}
\title{Download all clinical trials via specified APIv2 ULR}
\usage{
download_v2_url(url, folder, index = 1)
}
\arguments{
\item{url}{A string URL for the ClinicalTrials.gov APIv2}

\item{folder}{Folder to save download file with trials (before concatenation)}

\item{index}{Index value for naming downloaded files}
}
\value{
0 if the number of trials is 0 or the number of trials. The function primarily
saves data in the specified folder
Properties:
* Sensitive to provided URL. Some URLs may not be supported by https://clinicaltrials.gov/data-api/api
}
\description{
`download_v2_url()` is used to download trials for
a provided APIv2 URL (see more: https://clinicaltrials.gov/data-api/api). Currently
it can download 1000 trials per request. If resulting number of trials > 1000,
more than one request URL will be generated internally
}
\details{
Could be used for any URL supported by https://clinicaltrials.gov/data-api/api
Primarily intended for internal use
}
\examples{

download_v2_url(url = "https://clinicaltrials.gov/api/v2/studies?format=csv&query.cond=lung+cancer&query.intr=&pageSize=1000",
folder = "test2",
index = 3)
}
