% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_new_api.R
\name{get_trials_API_query}
\alias{get_trials_API_query}
\title{Get 1-100 trials from a specified query to the ClinicalTrials.gov API}
\usage{
get_trials_API_query(query, download_folder = NULL, min_rnk = 1, max_rnk = 100)
}
\arguments{
\item{query}{A string URL for the ClinicalTrials.gov API}

\item{download_folder}{Optional. Folder to save JSON file with trials (before concatenation)}

\item{min_rnk}{Starting index of downloaded trials}

\item{max_rnk}{Ending index of downloaded trials (max_rnk - min_rnk must be <= 100)}
}
\value{
A data.frame with clinical trials. Missing columns are handled via [data.table::rbindlist()].
Could return NA (if 0 trials)
Properties:
* May contain big number of columns
}
\description{
`get_trials_API_query()` used to download limited number of studies were the
number is not more than 100 trials. Currently only "full study" format is supported.
Look here for detail: https://clinicaltrials.gov/api/gui/home
}
\details{
The function is not useful for big lists of clinical trials!
Use [clinicaltrialsgov::prepare_download_URLs()] and
[clinicaltrialsgov::get_trials_API_urls()] instead
}
\examples{

get_trials_API_query(query = "lung+cancer", download_folder = "test")
}
