% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_new_api.R
\name{get_trials_API_urls}
\alias{get_trials_API_urls}
\title{Download all trials from vector of URLs to ClinicalTrials.gov API}
\usage{
get_trials_API_urls(url_list, delay = 0.5, download_folder = NULL)
}
\arguments{
\item{url_list}{A character vector of urls from [clinicaltrialsgov::prepare_download_URLs()]}

\item{delay}{A delay in seconds between requests}

\item{download_folder}{Optional. Folder to save JSON file with trials (before concatenation)}
}
\value{
A data.frame with clinical trials. Missing columns are handled via [data.table::rbindlist()].
}
\description{
`get_trials_API_urls()` This function is used to download all clinical trials from a character vector
of ClinicalTrials.gov API URLs.
Currently only "full study" format is supported.
Note: execution could take relatively long time depending on the number of trials. The resulting
data.frame may have a lot of columns (thousands)
Look here for detail: https://clinicaltrials.gov/api/gui/home
}
\details{
The function download all trials and concatenates the output. If download fails, it tries to download file 10 times and then proceeds to
next URL. A folder could be specified to stode all trials in the JSON format (before concatenation)
}
\examples{

urls <- prepare_download_URLs(query = "lung+cancer")
trials <- get_trials_API_urls(url_list = urls, download_folder = "test")
}
