% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_v2_api.R
\name{import_ct_results_v2_csv}
\alias{import_ct_results_v2_csv}
\title{Import downloaded trials into R}
\usage{
import_ct_results_v2_csv(folder)
}
\arguments{
\item{folder}{A folder with stored downloaded files via clinical_trial_downloaderV2_API()}
}
\value{
Data frame
}
\description{
`import_ct_results_v2_csv()` is to import all downloaded trials from a specified folder
}
\details{
Currently supports only .csv format of downloads.
}
\examples{

clinical_trial_downloaderV2_API(condition_terms = c("brain cancer", "glioma", "lung cancer"),
                               treatment_terms = c("cisplatin", "temozolomide", "olaparib"),
                               folder = "test_downloads3",
                               url_suffix = "&query.patient=AREA[LocationCity] London")
results = import_ct_results_v2_csv("test_downloads3")
}
