% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_v2_api.R
\name{list_to_df}
\alias{list_to_df}
\title{A function to concatenate list of data frames into a single data frame}
\usage{
list_to_df(data_list)
}
\arguments{
\item{data_list}{A list of data frames}
}
\value{
A data frame
Properties:
Each data frame in initial list must have the same column names
}
\description{
`list_to_df()` This function is used internally
}
\details{
Intended for internal use
}
\examples{
df1 = data.frame(
 y1 = c(1, 2, 3),
 y2 = c(4, 5, 6)
)
# Create another dataframe
df2 = data.frame(
 y1 = c(7, 8, 9),
 y2 = c(1, 4, 6)
)
# Create list of data frame using list()
listOfDataframe = list(df1, df2)
concatenated_list = list_to_df(listOfDataframe)
}
