% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_new_api.R
\name{parse_single_trial_recursive}
\alias{parse_single_trial_recursive}
\title{Recursive parsing of a clinical trial XML record from ClinicalTrials.gov}
\usage{
parse_single_trial_recursive(x, previous_name = NULL)
}
\arguments{
\item{x}{A clincal trial record in the list format obtained directly from [xml2::as_list()]}

\item{previous_name}{Typically is null and changed only within recursion}
}
\value{
A list representation of a clinical trial record.
Properties:
* Returned lists eventually terminate with a character vector with "Name" and "Value"
}
\description{
`parse_single_trial_recursive()` This function is used internally in other functions to parse
XML response for a clinical trial.
}
\details{
This is a recursive function. It is intebded for internal use.
}
\examples{
url <- "https://ClinicalTrials.gov/api/query/full_studies?expr=olaparib&min_rnk=1&max_rnk=100"
data <- httr::GET(url)
data_parsed <- xml2::read_xml(data)
trial_list <- xml2::xml_find_all(data_parsed, "FullStudyList")
trial_list <- xml2::as_list(trial_list)
trial_list <- trial_list[[1]]
single_trial <- trial_list[[1]]
single_trial <- parse_single_trial_recursive(single_trial)
}
