% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_new_api.R
\name{prepare_download_URLs}
\alias{prepare_download_URLs}
\title{Prepare download URLs to get all trials for the specified query to the ClinicalTrials.gov API}
\usage{
prepare_download_URLs(query)
}
\arguments{
\item{query}{A string URL for ClinicalTrials.gov API}
}
\value{
A character vector of ready-to-use URLs for the ClinicalTrials.gov API or NA (if 0 trials)
Properties:
* Each URL corresponds to 100 trials
}
\description{
`prepare_download_URLs()` used to prepare a character vector of URLs to subsequently pass to
[clinicaltrialsgov::get_trials_API_urls()]
Currently only "full study" format is supported.
Look here for detail: https://clinicaltrials.gov/api/gui/home
}
\details{
The function just prepares URLs for [clinicaltrialsgov::get_trials_API_urls()]
}
\examples{

prepare_download_URLs(query = "lung+cancer")
}
