% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts_legacy_api.R
\name{valid_url}
\alias{valid_url}
\title{Checks if the requested URL exists}
\usage{
valid_url(url, t = 2)
}
\arguments{
\item{url}{A string URL to check}

\item{t}{Timeout in seconds for connection}
}
\value{
Boolean. TRUE if the URL is valid, and FALSE if it is not
}
\description{
`valid_url()` This function is used to check if the requested URL is valid. This function is
expected to be used internally
}
\details{
This function is used to check if the requested URL is valid. This function is
expected to be used internally. It tries to establish conntection to the requested URL
}
\examples{

valid_url(url = "https://cran.r-project.org/")
valid_url(url = "https://i-dont-exist-at-all.com/")
}
